ADDED
This Restaurant Management System Project in Django created based on python, Django, and SQLITE3 Database. Restaurant Management System Project created using Python Django Framework on the backend and HTML, CSS, and JavaScript on the frontend. It has Admin side which allows extra features like manage food, sales, order, users, and inventory.

We present a web-based restaurant management framework for managing food orders for a given menu and a given customer in this project. It also offers home delivery services. It also has an admin interface that allows you to see and access all of the users, their orders, and the Restaurant’s menu.